refactor: hunt down and kill to_list's evil twin, typos
